building_health/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.0.0
opencv-python==4.11.0.86
pillow==10.4.0
fpdf==1.7.2
cloudinary==1.44.1
pymongo==4.6.0
openai==0.28.0
google-generativeai==0.8.5
ultralytics==8.0.0
